src/modules/orders/index.ts
import type { Cell } from '@ckb-lumos/base'
import { injectable } from 'inversify'
import { getConnection } from 'typeorm'
import OrderRepository from './order.repository'
import { OrderType } from './order.entity'
import { OrderDto } from './order.dto'
import DealRepository from './deal.repository'
import { parseOrderCell } from '../../utils'
import { Deal, DealStatus } from './deal.entity'

// const logTag = `\x1b[35m[Orders Service]\x1b[0m`

@injectable()
class OrdersService {
  #orderRepository = getConnection(process.env.NODE_ENV).getCustomRepository(OrderRepository)
  #dealRepository = getConnection(process.env.NODE_ENV).getCustomRepository(DealRepository)

  public saveOrder = (cell: Cell) => {
    const parsed = parseOrderCell(cell)
    return this.#orderRepository.saveOrder(parsed)
  }

  public removeOrder = (id: string) => {
    return this.#orderRepository.removeOrder(id)
  }

  /**
   * @param pageNo start from 0
   */
  public getAskOrders = async (args: string, pageNo = 0): Promise<OrderDto[]> => {
    const pendingOrderIds = await this.#dealRepository.getPendingOrderIds(args)
    return this.#orderRepository.getOrders(pageNo, OrderType.Ask, pendingOrderIds, args)
  }

  /**
   * @param pageNo start from 0
   */
  public getBidOrders = async (args: string, pageNo = 0): Promise<OrderDto[]> => {
    const pendingOrderIds = await this.#dealRepository.getPendingOrderIds(args)
    return this.#orderRepository.getOrders(pageNo, OrderType.Bid, pendingOrderIds, args)
  }

  public flushOrders = (cells: Array<Cell>) => {
    return this.#orderRepository.flushAllOrders(cells.map(parseOrderCell))
  }

  public clearOrders = () => {
    return this.#orderRepository.clear()
  }

  public saveDeal = (deal: Omit<Deal, 'createdAt'>) => {
    return this.#dealRepository.saveDeal(deal)
  }

  public updateDealStatus = (txHash: string, status: DealStatus) => {
    return this.#dealRepository.updateDealStatus(txHash, status)
  }

  public removeDeal = (txHash: string) => {
    return this.#dealRepository.removeDeal(txHash)
  }

  public getDeals = (pageNo: number, args: string) => {
    return this.#dealRepository.getDeals(pageNo, args)
  }

  public getPendingDeals = () => {
    return this.#dealRepository.getPendingDeals()
  }
}

export default OrdersService
